Issue #20455: math.ceil() returns an int in Python 3, no need to cast the result again to int

+++ Lib/selectors.py
@@ -372,7 +372,7 @@ if hasattr(select, 'poll'):
             else:
                 # poll() has a resolution of 1 millisecond, round away from
                 # zero to wait *at least* timeout seconds.
-                timeout = int(math.ceil(timeout * 1e3))
+                timeout = math.ceil(timeout * 1e3)
             ready = []
             try:
                 fd_event_list = self._poll.poll(timeout)
